Eliminate a few more undefined variables
svn path=/trunk/boinc/; revision=8401

+++ html/user/delete_profile.php
@@ -6,7 +6,7 @@ require_once("../inc/profile.inc");
 db_init();
 $user = get_logged_in_user();
 
-if ($_POST['delete']) {
+if (isset($_POST['delete']) && $_POST['delete']) {
     delete_profile($user);
     exit();
 }
